==> 03_leaf_correlations.Rmd <==
---
title: "LEAF 2021 dotazníkové šetření"
subtitle: "Přehled korelací"
team: výzkumný tým SCHOLA EMPIRICA # delete whole entry if undesirable
author:
- "Aleš Vomáčka^[Kontaktní autor. E-mail: vomacka@scholaempirica.org]" # complex entry must be quoted
- Martina Koutná
- Marek Havrda # "and"/"a" is added automatically in selected locale
date: "`r reschola::as_czech_date(Sys.time())`" # if Czech, use reschola::as_czech_date(Sys.time())
output: 
  reschola::schola_pdf:
    num_format: cs # inline formatting for numbers (1 000,00 with "cs")
    number_sections: true
toc-title: "Obsah"
abstract-title: Abstract
lang: cs # smart-quotes and author variable depends on the lang variable
logo: logos/schola # chars which are escaped by pandoc (i.e. "_") are prohibited
logo_height: 20pt
logo_client: false
logo_client_height: 30pt
head_height: false # may be used if fancyhdr complains 
---

```{r setup, include=FALSE}
options(scipen = 10) # force R to show full numbers, not scientific notation

library(reschola)
library(tidyverse)
library(scales)
library(kableExtra)
library(flextable)
library(gtsummary)
library(patchwork)
set_reschola_ggplot_fonts() # make ggplot2 use Roboto fonts without you having to set it
theme_set(theme_schola()) # default ggplot theme
source("98_custom_functions.R", encoding = "UTF-8")
source("shared.R", local = TRUE, encoding = "UTF-8") # helper functions and shared code
```

```{r data}
leaf <- read_rds("data-processed/leaf.rds")
```

```{r var-labels}
# All of this is the preparation of the variable labels data frame
# the data frame has to have two variables, "name" and "label" for desc_table() to function

var_labs <- attr(leaf, "variable.labels")
var_labs <- if_else(str_detect(var_labs, "\\]"), str_extract(var_labs, "(^.+)\\]"), var_labs)
var_labs <- str_replace(var_labs, "\\[", "")
var_labs <- str_replace(var_labs, "\\]", "")
var_labs <- str_trim(var_labs)

var_labs_df <- tibble(name  = names(leaf),
                      label = var_labs)

var_labs_df$label <- case_when(var_labs_df$name == "S4Q1" ~ "Pohlaví",
                               var_labs_df$name == "S4Q2" ~ "Věk",
                               var_labs_df$name == "S4Q3" ~ "Dosažené vzdělání",
                               var_labs_df$name == "S4Q7" ~ "Ekonomický profil",
                               var_labs_df$name == "S5Q2" ~ "Souhlasí s rozhovorem?",
                               var_labs_df$name == "S2Q3_a" ~ "LEAF ovplyvnil moju životnú cestu zásadným pozitívnym spôsobom\n(1 = určite nie, 6 = určite áno)",
                               var_labs_df$name == "S5Q1_a" ~ "môj vplyv na spoločenské dianie\n(1 = veľmi veľký vplyv, 5 = nulový vplyv)",
                               TRUE ~ var_labs_df$label)
```

# Postoje k současnostni a budoucnosti Slovenska podle programu

```{r future}
# quick and dirty function to generate results per absolved program
future_per_program <- function(future_var) {
leaf %>% 
  mutate(across(starts_with("S1Q1"),
                ~fct_relevel(., "neviem posúdiť")),
         across(starts_with("S1Q1"),
                ~fct_rev(.))) %>% 
  select((contains("S4Q4") & !contains("comment") | {{ future_var }})) %>% 
  pivot_longer(-{{ future_var }}) %>% 
  filter(value == "Áno") %>%
  left_join(., var_labs_df, by = "name") %>% 
    select(-c("value", "name")) %>% 
  tbl_summary(by = {{ future_var }}, percent = "row",
              label = list(label ~ "Absolvovaný kurz")) %>% 
  modify_header(label = "") %>%
    modify_footnote(update = everything() ~ NA) %>%
    bold_labels() %>%
    as_flex_table() %>%
    theme_booktabs() %>%
    FitFlextableToPage()
}
```

Otázka: *Na úvod, priblížme si realitu, v ktorej žijeme. Zaujíma nás tvoje osobné stanovisko k nasledovným výrokom.*

V tabulkách jsou řádková procenta

## Aktuálne nemá zmysel politicky sa angažovať na Slovensku.

```{r future-a}
future_per_program("S1Q1_a")
```

\newpage

## Vyrušuje ma, keď sa ľudia občiansky neangažujú.

```{r future-b}
future_per_program("S1Q1_b")
```

\newpage

## Úspech v dnešnom svete vyžaduje ťažké etické kompromisy.

```{r future-c}
future_per_program("S1Q1_c")
```

\newpage

## Ak chce človek niečo zmeniť, je to najmä v jeho rukách.

```{r future-d}
future_per_program("S1Q1_d")
```

\newpage

## Pre udržanie kvalitnej a dobre platenej práce musí byt človek flexibilný a pracovať i pre ľudí s neinšpiratívnymi hodnotami.

```{r future-e}
future_per_program("S1Q1_e")
```

\newpage

## Najbližších 10 rokov je na Slovensku pre mňa cesta zarúbaná.

```{r future-f}
future_per_program("S1Q1_f")
```

\newpage

# Vlivné sociální skupiny podle programu

```{r social groups}
# quick and dirty function to generate results per absolved program
social_per_program <- function(social_var) {
leaf %>% 
  select((contains("S4Q4") & !contains("comment") | {{ social_var }})) %>% 
  pivot_longer(-c( {{ social_var }} )) %>% 
  filter(value == "Áno") %>%
  left_join(., var_labs_df, by = "name") %>% 
    select(-c("value", "name")) %>% 
  group_by(label) %>% 
  summarise(`průměr` = mean({{ social_var }}),
            `průměr` = round(`průměr`),
            `počet odpovědí` = n()) %>% 
  rename("Organizace" = "label") %>% 
  flextable() %>%
  theme_booktabs() %>%
  FitFlextableToPage() %>% 
  footnote(j = 2,
           part = "header",
           value = as_paragraph("Stupnice od 0 do 100"))
}
```

## rodičia, príbuzní

Otázka: **Nakoľko dôležité boli pri rozhodovaní sa o smerovaní tvojej študijnej / pracovnej* dráhy nasledujúci ľudia či komunity, ktorými si bol(a) obklopený(-á)?**

```{r social-family}
social_per_program(S2Q1_a)
```

\newpage

## Priateľ(-ka), kamaráti, známi

Otázka: **Nakoľko dôležité boli pri rozhodovaní sa o smerovaní tvojej študijnej / pracovnej* dráhy nasledujúci ľudia či komunity, ktorými si bol(a) obklopený(-á)?**

```{r friends}
social_per_program(S2Q1_b)
```

\newpage

## Učiteľ/učiteľka v škole, mentor

Otázka: **Nakoľko dôležité boli pri rozhodovaní sa o smerovaní tvojej študijnej / pracovnej* dráhy nasledujúci ľudia či komunity, ktorými si bol(a) obklopený(-á)?**

```{r teachers}
social_per_program(S2Q1_c)
```

\newpage

## Nezisková organizácia, občianske združenie

Otázka: **Nakoľko dôležité boli pri rozhodovaní sa o smerovaní tvojej študijnej / pracovnej* dráhy nasledujúci ľudia či komunity, ktorými si bol(a) obklopený(-á)?**

```{r nonprofit}
social_per_program(S2Q1_d)
```

\newpage

## Náboženská komunita a ľudia v nej

Otázka: **Nakoľko dôležité boli pri rozhodovaní sa o smerovaní tvojej študijnej / pracovnej* dráhy nasledujúci ľudia či komunity, ktorými si bol(a) obklopený(-á)?**

```{r religion}
social_per_program(S2Q1_e)
```

\newpage

## Inšpiratívny človek, konkrétna osobnosť,  odborník(-čka

Otázka: **Nakoľko dôležité boli pri rozhodovaní sa o smerovaní tvojej študijnej / pracovnej* dráhy nasledujúci ľudia či komunity, ktorými si bol(a) obklopený(-á)?**

```{r person}
social_per_program(S2Q1_f)
```

\newpage

## Voľnočasová komunita

Otázka: **Nakoľko dôležité boli pri rozhodovaní sa o smerovaní tvojej študijnej / pracovnej* dráhy nasledujúci ľudia či komunity, ktorými si bol(a) obklopený(-á)?**

```{r hobby}
social_per_program(S2Q1_g)
```

\newpage

## Iné

Otázka: **Nakoľko dôležité boli pri rozhodovaní sa o smerovaní tvojej študijnej / pracovnej* dráhy nasledujúci ľudia či komunity, ktorými si bol(a) obklopený(-á)?**

```{r other}
social_per_program(S2Q1_other)
```

\newpage

# Vnímaný pozitivní vliv LEAFu podle programu

Otázka: **LEAF ovplyvnil moju životnú cestu zásadným pozitívnym spôsobom (1 = určite nie, 6 = určite áno)**

V tabulce jsou řádková procenta

```{r leaf-impact-program}
leaf %>% 
  select((contains("S4Q4") & !contains("comment") | S2Q3_a)) %>% 
  pivot_longer(-S2Q3_a) %>% 
  filter(value == "Áno") %>%
  left_join(., var_labs_df, by = "name") %>% 
    select(-c("value", "name")) %>% 
  tbl_summary(by = S2Q3_a, percent = "row",
              label = list(label ~ "Absolvovaný kurz")) %>% 
  modify_header(label = "") %>%
    modify_footnote(update = everything() ~ NA) %>%
    bold_labels() %>%
    as_flex_table() %>%
    theme_booktabs() %>%
    FitFlextableToPage()
```

\newpage

# Současnost a budoucnost Slovenska podle vlivu LEAFu

Otázka 1: **Na úvod, priblížme si realitu, v ktorej žijeme. Zaujíma nás tvoje osobné stanovisko k nasledovným výrokom.**
Otázka 2: **LEAF ovplyvnil moju životnú cestu zásadným pozitívnym spôsobom (1=určite nie; 6=určite áno)**

```{r future-impact-a}
leaf %>% 
  mutate(S2Q3_a = case_when(S2Q3_a >= 5 ~ "5-6",
                            S2Q3_a >= 3 ~ "3-4",
                            S2Q3_a >= 1 ~ "1-2",
                            TRUE        ~ NA_character_)) %>% 
  desc_table(vars = c(S1Q1_a, S2Q3_a), var_labs_df, groupby = "S2Q3_a")
```


```{r future-impact-b}
leaf %>% 
  mutate(S2Q3_a = case_when(S2Q3_a >= 5 ~ "5-6",
                            S2Q3_a >= 3 ~ "3-4",
                            S2Q3_a >= 1 ~ "1-2",
                            TRUE        ~ NA_character_)) %>% 
  desc_table(vars = c(S1Q1_b, S2Q3_a), var_labs_df, groupby = "S2Q3_a")
```

```{r future-impact-c}
leaf %>% 
  mutate(S2Q3_a = case_when(S2Q3_a >= 5 ~ "5-6",
                            S2Q3_a >= 3 ~ "3-4",
                            S2Q3_a >= 1 ~ "1-2",
                            TRUE        ~ NA_character_)) %>% 
  desc_table(vars = c(S1Q1_c, S2Q3_a), var_labs_df, groupby = "S2Q3_a")
```

```{r future-impact-d}
leaf %>% 
  mutate(S2Q3_a = case_when(S2Q3_a >= 5 ~ "5-6",
                            S2Q3_a >= 3 ~ "3-4",
                            S2Q3_a >= 1 ~ "1-2",
                            TRUE        ~ NA_character_)) %>% 
  desc_table(vars = c(S1Q1_d, S2Q3_a), var_labs_df, groupby = "S2Q3_a")
```

```{r future-impact-e}
leaf %>% 
  mutate(S2Q3_a = case_when(S2Q3_a >= 5 ~ "5-6",
                            S2Q3_a >= 3 ~ "3-4",
                            S2Q3_a >= 1 ~ "1-2",
                            TRUE        ~ NA_character_)) %>% 
  desc_table(vars = c(S1Q1_e, S2Q3_a), var_labs_df, groupby = "S2Q3_a")
```

```{r future-impact-f}
leaf %>% 
  mutate(S2Q3_a = case_when(S2Q3_a >= 5 ~ "5-6",
                            S2Q3_a >= 3 ~ "3-4",
                            S2Q3_a >= 1 ~ "1-2",
                            TRUE        ~ NA_character_)) %>% 
  desc_table(vars = c(S1Q1_f, S2Q3_a), var_labs_df, groupby = "S2Q3_a")
```


# Rozložení dimenzí podle vlivu LEAFu

## Dimenze 1

Dimenze 1 je ekvivalentní otázce na vnímaný vliv LEAFu: **LEAF ovplyvnil moju životnú cestu zásadným pozitívnym spôsobom.**

```{r dimension1}
leaf %>% 
  count(dimension1) %>% 
  ggplot(aes(x = dimension1, y = n)) +
  geom_col() +
  scale_x_continuous(breaks = seq(1, 6, 1)) +
  theme_schola(axis.title = element_text()) +
  labs(x = "Dimenze 1", y = "Počet respondentů")
```

## Dimenze 2

Grafy tříděny podle výroku: **LEAF ovplyvnil moju životnú cestu zásadným pozitívnym spôsobom (1=určite nie; 6=určite áno)**

```{r dimension2}
leaf %>% 
  filter(!is.na(S2Q3_a)) %>% 
  count(dimension2, S2Q3_a) %>%
  group_by(S2Q3_a) %>% 
  mutate(freq = n/sum(n),
         freq = percent(freq, 1),
         S2Q3_a = case_when(S2Q3_a == 1 ~ "1=určite nie",
                            S2Q3_a == 6 ~ "6=určite áno",
                            TRUE        ~ as.character(S2Q3_a))) %>% 
  ggplot(aes(x = dimension2, y = n, label = freq)) +
  geom_text(nudge_y = 2) +
  geom_col() +
  facet_wrap(~S2Q3_a) +
  theme_schola(multiplot = TRUE)
```

## Dimenze 3

Grafy tříděny podle výroku: **LEAF ovplyvnil moju životnú cestu zásadným pozitívnym spôsobom (1=určite nie; 6=určite áno)**

```{r dimension3}
leaf %>% 
  filter(!is.na(S2Q3_a)) %>% 
  mutate(S2Q3_a = case_when(S2Q3_a == 1 ~ "1=určite nie",
                            S2Q3_a == 6 ~ "6=určite áno",
                            TRUE        ~ as.character(S2Q3_a))) %>% 
  ggplot(aes(x = dimension3)) +
  geom_histogram(bins = 15, color = "white") +
  facet_wrap(~S2Q3_a) +
  theme_schola(multiplot = TRUE)
```

## Dimenze 4

Grafy tříděny podle výroku: **LEAF ovplyvnil moju životnú cestu zásadným pozitívnym spôsobom (1=určite nie; 6=určite áno)**

```{r dimension4}
leaf %>% 
  filter(!is.na(S2Q3_a)) %>% 
  count(dimension4, S2Q3_a) %>% 
  group_by(S2Q3_a) %>% 
  mutate(S2Q3_a = case_when(S2Q3_a == 1 ~ "1=určite nie",
                            S2Q3_a == 6 ~ "6=určite áno",
                            TRUE        ~ as.character(S2Q3_a)),
         freq = n/sum(n),
         freq = percent(freq, 1)) %>% 
  ggplot(aes(x = dimension4, y = n, label = freq)) +
  geom_col() +
  scale_x_continuous(breaks = seq(0,10,2)) +
  facet_wrap(~S2Q3_a) +
  theme_schola(multiplot = TRUE)
```

\newpage

# Bodové grafy/matice dimenzí

Každá kombinace dimenzí je zobrazena pouze jednou.

## Dimeze 1 vůči ostatním dimenzím

```{r d1d2, fig.cap= "Dimenze 1 a Dimenze 2"}
leaf %>% 
  ggplot(aes(x = dimension1, y = dimension2, group = dimension1)) +
  geom_jitter(width = 0.2, height = 0) +
  stat_summary(fun = mean,
               geom = "crossbar",
               mapping = aes(color = "Průměr skupiny")) +
  scale_color_manual(values = "red") +
  scale_x_continuous(breaks = 1:6) +
  labs(x = "Dimenze 1", y = "Dimenze 2", color = element_blank()) +
  theme_schola(axis.title = element_text()) +
  theme(legend.position = "bottom")

```

```{r d1d3, fig.cap= "Dimenze 1 a Dimenze 3"}
leaf %>% 
  ggplot(aes(x = dimension1, y = dimension3, group = dimension1)) +
  geom_jitter(width = 0.2, height = 0) +
  stat_summary(fun = mean,
               geom = "crossbar",
               mapping = aes(color = "Průměr skupiny")) +
  scale_color_manual(values = "red") +
  scale_x_continuous(breaks = 1:6) +
  labs(x = "Dimenze 1", y = "Dimenze 3", color = element_blank()) +
  theme_schola(axis.title = element_text()) +
  theme(legend.position = "bottom")
```

```{r d1d4, fig.cap= "Dimenze 1 a Dimenze 4"}
leaf %>% 
  ggplot(aes(x = dimension1, y = dimension4, group = dimension1)) +
  geom_jitter(width = 0.2, height = 0) +
  stat_summary(fun = mean,
               geom = "crossbar",
               mapping = aes(color = "Průměr skupiny")) +
  scale_color_manual(values = "red") +
  scale_x_continuous(breaks = 1:6) +
  labs(x = "Dimenze 1", y = "Dimenze 4", color = element_blank()) +
  theme_schola(axis.title = element_text()) +
  theme(legend.position = "bottom")
```

\newpage

## Dimenze 2 vůči ostatním dimenzím

```{r d2d3, fig.cap="Dimenze 2 a Dimenze 3"}
leaf %>% 
  ggplot(aes(x = dimension2, y = dimension3, group = dimension2)) +
  geom_jitter(width = 0.2, height = 0) +
  stat_summary(fun = mean,
               geom = "crossbar",
               mapping = aes(color = "Průměr skupiny")) +
  scale_color_manual(values = "red") +
  scale_x_continuous(breaks = 1:6) +
  labs(x = "Dimenze 2", y = "Dimenze 3", color = element_blank()) +
  theme_schola(axis.title = element_text()) +
  theme(legend.position = "bottom")
```

```{r d2d4, fig.cap="Dimenze 2 a Dimenze 4"}
leaf %>% 
  ggplot(aes(x = dimension2, y = dimension4, group = dimension2)) +
  geom_jitter(width = 0.2, height = 0) +
  stat_summary(fun = mean,
               geom = "crossbar",
               mapping = aes(color = "Průměr skupiny")) +
  scale_color_manual(values = "red") +
  scale_x_continuous(breaks = 1:6) +
  labs(x = "Dimenze 2", y = "Dimenze 4", color = element_blank()) +
  theme_schola(axis.title = element_text()) +
  theme(legend.position = "bottom")
```

## Dimenze 3 vůči ostatním

```{r d3d4, fig.cap="Dimenze 3 a Dimenze 4"}
leaf %>% 
  ggplot(aes(x = dimension3, y = dimension4, group = dimension3)) +
  geom_count() +
  labs(x = "Dimenze 3", y = "Dimenze 4", size = "Počet respondentů") +
  theme_schola(axis.title = element_text()) +
  theme(legend.position = "bottom")
```


# Korelace mezi vnímaným vlivem LEAFu a současnosti/budoucnosti Slovenska

Otázka na přítomnost / budoucnost Slovenska: **Na úvod, priblížme si realitu, v ktorej žijeme. Zaujíma nás tvoje osobné stanovisko k nasledovným výrokom. (škála vôbec nesúhlasím - určite súhlasím).**


Otázka na přínos LEAFu: **Čo myslíš, do akej miery ti LEAF pomohol v nasledujúcich oblastiach osobného alebo profesijného života? (škála vôbec - zásadne).**

Korelace měřena pomocí Kendallova tau (0 = žádná korelace, 1 = dokonalá přímá úměra, -1 = dokonalá nepřímá úměra).

```{r impact-future-corr}
cor_df <- leaf %>% 
            select(starts_with("S2Q5") | starts_with("S1Q1")) %>% 
            mutate(across(.cols = starts_with("S2Q5"),
                          .fns  = ~na_if(., "netýka sa ma")),
                   across(.cols = starts_with("S1Q1"),
                          .fns  = ~na_if(.,"neviem posúdiť")),
                   across(.cols = everything(),
                          .fns  = droplevels),
                   across(.cols = everything(),
                          .fns  = as.numeric))

cor_matrix <-  round( cor(cor_df, method = "kendall", use = "complete.obs"),2 )

cor_df <-  as.data.frame(cor_matrix)

cor_df <- rownames_to_column(cor_df, var = "var1")
cor_df <- pivot_longer(cor_df,
                       cols = -var1,
                       names_to = "var2",
                       values_to = "corr")
cor_df %>% 
  filter(str_detect(var1, "S1Q1") & !str_detect(var2, "S1Q1")) %>% 
  rename("name" = "var1") %>% 
  left_join(., var_labs_df, by = "name") %>% 
  select(-name) %>% 
  rename("name" = "var2",
         "var1" = "label") %>% 
  left_join(., var_labs_df, by = "name") %>% 
  select(-name) %>% 
  rename("var2" = "label") %>% 
  relocate(corr, .after = last_col()) %>% 
  flextable() %>% 
  theme_booktabs() %>% 
  FitFlextableToPage() %>% 
  width(j = 1, 2) %>% 
  width(j = 2, 2) %>% 
  set_header_labels(var1 = "Současnost / budoucnost Slovenska",
                    var2 = "Přínos LEAFu",
                    corr = "Kendall tau") %>% 
  fontsize(j = 3, size = 8)
```

\newpage

# Korelace mezi vnímaným vlivem LEAFu a aktivitami za poslední 2 roky

Otázka na přítomnost / budoucnost Slovenska: **Na úvod, priblížme si realitu, v ktorej žijeme. Zaujíma nás tvoje osobné stanovisko k nasledovným výrokom. (škála vôbec nesúhlasím - určite súhlasím).**

Otázka na aktivity za poslední 2 roky: **Čomu z nasledujúcich aktivít sa za posledný rok-dva systematicky venuješ? ( škála nie - áno)**

Korelace měřena pomocí Kendallova tau (0 = žádná korelace, 1 = dokonalá přímá úměra, -1 = dokonalá nepřímá úměra).

```{r impact-activities-corr}
cor_df <- leaf %>% 
            select(starts_with("S3Q1") | starts_with("S1Q1")) %>% 
            mutate(across(.cols = starts_with("S3Q1"),
                          .fns  = ~na_if(., "netýka sa ma")),
                   across(.cols = starts_with("S3Q1"),
                          .fns  = ~fct_relevel(., "nie", "áno")),
                   across(.cols = starts_with("S1Q1"),
                          .fns  = ~na_if(.,"neviem posúdiť")),
                   across(.cols = everything(),
                          .fns  = droplevels),
                   across(.cols = everything(),
                          .fns  = as.numeric))

cor_matrix <-  round( cor(cor_df, method = "kendall", use = "complete.obs"),2 )

cor_df <-  as.data.frame(cor_matrix)

cor_df <- rownames_to_column(cor_df, var = "var1")
cor_df <- pivot_longer(cor_df,
                       cols = -var1,
                       names_to = "var2",
                       values_to = "corr")
cor_df %>% 
  filter(str_detect(var1, "S1Q1") & !str_detect(var2, "S1Q1")) %>% 
  rename("name" = "var1") %>% 
  left_join(., var_labs_df, by = "name") %>% 
  select(-name) %>% 
  rename("name" = "var2",
         "var1" = "label") %>% 
  left_join(., var_labs_df, by = "name") %>% 
  select(-name) %>% 
  rename("var2" = "label") %>% 
  relocate(corr, .after = last_col()) %>% 
  flextable() %>% 
  theme_booktabs() %>% 
  FitFlextableToPage() %>% 
  width(j = 1, 2) %>% 
  width(j = 2, 2) %>% 
  set_header_labels(var1 = "Aktivita za poslední 2 roky",
                    var2 = "Přínos LEAFu",
                    corr = "Kendall tau") %>% 
  fontsize(j = 3, size = 8)
```

\newpage

# Vliv LEAFu a osobní vliv na veřejné dění

Znění Otázky na vliv LEAFu na životní cestu: **LEAF ovplyvnil moju životnú cestu zásadným pozitívnym spôsobom (1 = určite nie, 6 = určite áno)**

Znění otázky na osobní vliv na společenské dění: **Ako hodnotíš svoj celkový vplyv na spoločenské dianie okolo seba, ktorý máš zo svojej pozície? (1 = veľmi veľký vplyv, nulový vplyv)**

```{r impact-influnce-cor}
tbl_cross(leaf, row = S5Q1_a, col = S2Q3_a,
          label = list(S2Q3_a = "vplyv LEAF na životnú cestu",
                       S5Q1_a = "Vplyv na spoločenské dianie"),
          percent = "row",
          missing = "no",
          margin = NULL) %>% 
  modify_header(label = "") %>% 
  as_flex_table() %>% 
  theme_booktabs() %>% 
  FitFlextableToPage()
```
